Start splitting the app
