+++ components/experimental/diceroller-standalone/src/app/index.ts
@@ -2,6 +2,3 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-
-export * from "./container";
-export * from "./component";
